W3Oven 0.8b
Changed a lot:
-Added Workspace Manager where you can create, delete and load mods.
-The button "Create Mod or Add assets" allows you to recursively create
All the folders of the relative path of the asset. When the mod does not
exist the mod folder is created.
- Added a dropdown that lists just the folders inside the Workspace
directory that has a "Modded" folder to it
- Added a buttons to Open the mod folder and the Uncooked folder to help
adding assets manually.

- Now the textures ending with _d01, _n01, _s01, _e01, _a01 will have
texturegroup automatically set. Users can still set a custom
texturegroup.
- Added Auto Install checkboxes
When Enabling the Auto Install for the first time, user will be
requested to select the "The Witcher 3\Mods" folder.
If the Auto Install is enabled it will copy the packed mod folder there.
The checkbox Just texture.cache wiill make it copy just the
texture.cache file
The checkbox "Is Custom DLC" installs only the texture.cache to the
Custom DLC folder in "The Witcher 3\DLC"